Add AbstractHttpServlet class for common functionality in HTTP servlets
Created the AbstractHttpServlet class in the ru.starkov.servlet package to provide common functionality for HTTP servlets. This abstract class extends HttpServlet and is intended for initializing fields used in servlet subclasses. The init() method initializes the Gson object using the ServletContext attribute, allowing subclasses to use Gson for JSON serialization and deserialization.

.