Add CurrencyAlreadyExistException class for indicating existing currency
Created the CurrencyAlreadyExistException class in the ru.starkov.exception package to indicate that a currency already exists in the system. This exception is thrown when attempting to add a currency that is already present.

Key features of the CurrencyAlreadyExistException class include:

Extending the RuntimeException class for unchecked exception handling.
Providing a constructor to set the exception message.
This exception class enhances error handling in the application by providing a specific exception type for cases where attempting to add an existing currency occurs.